+++ inboxen/middleware.py
@@ -19,18 +19,12 @@
 
 from django.conf import settings
 from django.contrib import messages
-from django.core.exceptions import PermissionDenied
-from django.core.urlresolvers import reverse
 from django.shortcuts import redirect
 from django.utils.translation import ugettext as _
 
 from ratelimitbackend.exceptions import RateLimitException
-from sudo.views import redirect_to_sudo
-from wagtail.wagtailadmin.views.account import login, logout
 
 
-LOGIN_NAME = "{}.{}".format(login.__module__, login.__name__)
-LOGOUT_NAME = "{}.{}".format(logout.__module__, logout.__name__)
 SESSION_HALF_COOKIE_AGE = settings.SESSION_COOKIE_AGE / 2
 
 
@@ -56,33 +50,3 @@ class ExtendSessionMiddleware(object):
                 request.session.cycle_key()
                 request.session.set_expiry(settings.SESSION_COOKIE_AGE)
                 request.session.modified = True
-
-
-class RedirectWagLoginMiddleware(object):
-    """Redirects Wagtail's admin login/logout views
-
-    Neither handle 2FA or sudo cookies
-    """
-    def process_view(self, request, view_func, view_args, view_kwargs):
-        try:
-            view_path = "{}.{}".format(view_func.__module__, view_func.__name__)
-        except AttributeError:
-            view_path = ""
-
-        if view_path == LOGIN_NAME:
-            return redirect("user-login")
-        elif view_path == LOGOUT_NAME:
-            return redirect("user-logout")
-
-
-class WagtailAdminProtectionMiddleware(object):
-    """Protects Wagtail's admin with both sudo and 2FA"""
-    def process_request(self, request):
-        if request.path.startswith(str(settings.WAGTAIL_ADMIN_BASE_URL)):
-            if request.user.is_anonymous():
-                # admin has its own redirect for non-logged in users
-                return
-            elif not request.user.is_verified() and not settings.DEBUG:
-                raise PermissionDenied("Admins must have Two Factor Authentication enabled")
-            elif not request.is_sudo():
-                return redirect_to_sudo(request.get_full_path())